+++ tests/acceptance/test_user_behavior.py
@@ -140,7 +140,9 @@ def test_plotting_with_forced_ascii():
 def test_plotting_with_forced_ascii_and_custom_symbols():
     ys = [[1, 3, -2], [3, 4, 3, 4, 3, 5], [0]]
     symbols = ["A", "B", "C"]
-    strs = plot_to_string(ys, force_ascii=True, force_ascii_characters=symbols)
+    strs = plot_to_string(
+        ys, force_ascii=True, color=False, force_ascii_characters=symbols
+    )
     plot_str = "\n".join(strs)
 
     for i in range(2):